Docker update for rocBLAS, HCC, and HIP to 2.1 (#1606)
* Updated the requirements with newer HIP, HCC, rocBLAS, and clang-ocl

* Fixed boost issue.

* Fixed tidy issues across the lib.

* Formatting.

* Update hip and rocblas version.

* Fixed remaining tidy issues from formatting.

* Formatting.

* More tidy issues.

* Float conversion.

* NOLINT on stubborn coversion.

* ktime to float

* Remove unneeded NOLINT

* Removed unneeded semi-colon

* Drive updates.

* Switched to rocBLAS 2.3 RC commit hash.

* Use size_t for fusion op_map.


[ROCm/MIOpen commit: 81f7926ad69ccb837da9747f14dba7637e515966]

+++ src/solver/conv_ocl_dir2D_bwdWrW_2.cpp
@@ -694,7 +694,8 @@ ConvSolution ConvOclBwdWrW2<N_BATCH_LOOPS>::GetSolution(
         kernel.l_wk.push_back(1);
 
         assert(utility_read_unit != 0);
-        int gbl_ut_wk0 = wei_bstride * params.n_inputs / utility_read_unit;
+        int gbl_ut_wk0 =
+            static_cast<int>(static_cast<int>(wei_bstride) * params.n_inputs / utility_read_unit);
 
         kernel.g_wk.push_back(gbl_ut_wk0);
         kernel.g_wk.push_back(1);
